changed: put enums in separate file
to help break circular dependencies
